requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
faiss-cpu
sentence-transformers
pdfplumber
PyPDF2
python-multipart
orjson
